Move CustomObject Sub-Components Into Separate Packages
Move BusinessProcess, CustomField, FieldSet, RecordType, ValidationRule,
CompactLayout, ListView, SharingReason, and WebLink metadata into
separate packages in preparation for better source-format support.

Add CustomObjectTranslation and CustomFieldTranslation types.
